batch: Extract operation candidate types
Operation candidate construction now owns candidate enumeration, while
output and batch-source helpers still import preview records from that
construction module.

The project does not provide a lightweight import boundary for candidate
preview value types. Consumers that only format, persist, count, or
validate previews must depend on merge enumeration code they do not call.

This commit addresses that by moving candidate preview records, operation
literals, count results, and the enumeration-limit error into
batch.operation_candidate_types, then updating type-only consumers to
import that module.

Subsequent test commits will pin direct type imports and guard the
construction boundary.
