+++ src/git_stage_batch/commands/batch_source/candidate_materialization.py
@@ -7,9 +7,11 @@ from dataclasses import dataclass
 
 from . import candidate_inputs as _candidate_inputs
 from . import candidate_previews as _candidate_previews
-from ...batch.operation_candidates import (
+from ...batch.operation_candidate_types import (
     OperationCandidatePreview,
     TargetCandidatePreview,
+)
+from ...batch.operation_candidates import (
     build_apply_candidate_previews,
     build_include_candidate_previews,
 )
